pyproject.toml
[project]
name = "gargantua"
version = "0.1.0"
description = ""
authors = [
    {name = "deepak thapliyal",email = "deepak.t@mihira.studio"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "imath (>=0.0.2,<0.0.3)",
    "pandas (>=2.2.3,<3.0.0)",
    "ffmpeg-python (>=0.2.0,<0.3.0)",
]

[tool.poetry]
readme = ["README.md"]
packages = [
    {include = "gargantua", from = "src"},
]

[tool.poetry.scripts]
gargantua = "gargantua.main:main"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


